+++ lib/src/configuration/better_player_theme.dart
@@ -1,0 +1,1 @@
+enum BetterPlayerTheme { material, cupertino, custom }