+++ lib/src/controls/better_player_cupertino_controls.dart
@@ -3,10 +3,10 @@ import 'dart:async';
 import 'dart:ui' as ui;
 
 // Flutter imports:
+import 'package:better_player/src/configuration/better_player_controls_configuration.dart';
 import 'package:flutter/material.dart';
 
 // Project imports:
-import 'package:better_player/src/controls/better_player_controls_configuration.dart';
 import 'package:better_player/src/controls/better_player_controls_state.dart';
 import 'package:better_player/src/controls/better_player_cupertino_progress_bar.dart';
 import 'package:better_player/src/controls/better_player_progress_colors.dart';